fix(options): center directory clear button icon
